pyproject.toml
[tool.poetry]
name = "paperswithcode-notification"
version = "0.1.0"
description = ""
authors = ["Junya Morioka <mjun@mjunya.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
